TextAdventure/requirements.txt
pygame>=2.6.1
moderngl>=5.12.0
numpy>=2.3.3
pillow>=11.30